OPENDNSSEC-321 validate xml files on (re)read.

+++ enforcer-ng/src/enforcer/update_conf_cmd.cpp
@@ -42,48 +42,47 @@ static const char *module_str = "update_repositorylist_cmd";
 void
 help_update_conf_cmd(int sockfd)
 {
-    ods_printf(sockfd,
-			   "update repositorylist  Import respositories from conf.xml into the enforcer.\n");
+	ods_printf(sockfd,
+		"update repositorylist  Import respositories from conf.xml "
+		"into the enforcer.\n");
 }
 
 static void
 flush_all_tasks(int sockfd, engine_type* engine)
 {
-    ods_log_debug("[%s] flushing all tasks...", module_str);
-    ods_printf(sockfd,"flushing all tasks...\n");
+	ods_log_debug("[%s] flushing all tasks...", module_str);
+	ods_printf(sockfd,"flushing all tasks...\n");
 
-    ods_log_assert(engine);
-    ods_log_assert(engine->taskq);
-    lock_basic_lock(&engine->taskq->schedule_lock);
-    /* [LOCK] schedule */
-    schedule_flush(engine->taskq, TASK_NONE);
-    /* [UNLOCK] schedule */
-    lock_basic_unlock(&engine->taskq->schedule_lock);
-    engine_wakeup_workers(engine);
+	ods_log_assert(engine);
+	ods_log_assert(engine->taskq);
+	lock_basic_lock(&engine->taskq->schedule_lock);
+	/* [LOCK] schedule */
+	schedule_flush(engine->taskq, TASK_NONE);
+	/* [UNLOCK] schedule */
+	lock_basic_unlock(&engine->taskq->schedule_lock);
+	engine_wakeup_workers(engine);
 }
 
 int
-handled_update_conf_cmd(int sockfd, engine_type* engine, const char *cmd,
-		ssize_t n)
+handled_update_conf_cmd(int sockfd, engine_type* engine, 
+	const char *cmd, ssize_t n)
 {
-    const char *scmd = "update repositorylist";
+	const char *scmd = "update repositorylist";
 
-    cmd = ods_check_command(cmd,n,scmd);
-    if (!cmd)
-        return 0; // not handled
+	cmd = ods_check_command(cmd,n,scmd);
+	if (!cmd)
+		return 0; // not handled
 
-    ods_log_debug("[%s] %s command", module_str, scmd);
-    time_t tstart = time(NULL);
+	ods_log_debug("[%s] %s command", module_str, scmd);
+	time_t tstart = time(NULL);
 
-    perform_update_conf(engine,cmd,n);
+	if (perform_update_conf(sockfd, engine, cmd, n)) {
+		kill(engine->pid, SIGHUP);
+		ods_printf(sockfd, "Notifying enforcer of new respositories! \n");
+	}
 
-
-    kill(engine->pid, SIGHUP);
-    ods_printf(sockfd, "Notifying enforcer of new respositories! \n");
-
-
-    ods_printf(sockfd,"%s completed in %ld seconds.\n",scmd,time(NULL)-tstart);
-    return 1;
+	ods_printf(sockfd,"%s completed in %ld seconds.\n",scmd,time(NULL)-tstart);
+	return 1;
 }
